oblig ferdig

+++ Oblig1_AleksandraWos/Oppgave5.py
@@ -4,4 +4,5 @@ pers3 = 2.5
 pers4 = 21
 pers5 = 0
 
-print((pers1+pers2+pers3+pers4+pers5)/5)
+
+print(int((pers1+pers2+pers3+pers4+pers5)/5))
